Overhaul de-localization
General overhaul of the de-localization.

I feel like ```Speicherort``` as the main translation for ```storage``` is generally correct, but quite unusual. You should think about maybe changing it to something like ```Notizbuch = Notebook``` on many occasions. If you wish to implement that feel free to ask me to translate it.

'note.date': currently it says ```38 minutes her``` but it should either be ```38 Minuten her``` or ```vor 38 Minuten```. IMO the second one is better, but I think more code adjustment is necessary somewhere else to rephrase the wording.